Updates unit tests as per updates in email subject and body update
 renames __init__.py to test_notify.py
